+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2.2.1
 
+### Object Server API Changes (In Beta)
+
+* Fixed `SyncConfiguration.toString()` so it now outputs a correct description instead of an empty string (#3787).
+
 ### Bug fixes
 
 * Added version number to the native library, preventing ReLinker from accidentally loading old code (#3775).